+++ src/code42cli/cmds/users.py
@@ -1,14 +1,18 @@
 import click
 from pandas import DataFrame
 
+from code42cli.bulk import generate_template_cmd_factory
+from code42cli.bulk import run_bulk_process
 from code42cli.click_ext.groups import OrderedGroup
 from code42cli.click_ext.options import incompatible_with
 from code42cli.errors import Code42CLIError
 from code42cli.errors import UserDoesNotExistError
+from code42cli.file_readers import read_csv_arg
 from code42cli.options import format_option
 from code42cli.options import sdk_options
 from code42cli.output_formats import DataFrameOutputFormatter
 from code42cli.output_formats import OutputFormat
+from code42cli.output_formats import OutputFormatter
 
 
 @click.group(cls=OrderedGroup)
@@ -30,6 +34,11 @@ inactive_option = click.option(
     is_flag=True,
     help="Limits results to only deactivated users.",
     cls=incompatible_with("active"),
+)
+
+
+user_uid_option = click.option(
+    "--user-id", help="The unique identifier of the user to be modified.", required=True
 )
 
 
@@ -84,6 +93,95 @@ def remove_role(state, username, role_name):
     _remove_user_role(state.sdk, role_name, username)
 
 
+@users.command(name="update")
+@user_uid_option
+@click.option("--username", help="The new username for the user.")
+@click.option("--password", help="The new password for the user.")
+@click.option("--email", help="The new email for the user.")
+@click.option("--first-name", help="The new first name for the user.")
+@click.option("--last-name", help="The new last name for the user.")
+@click.option("--notes", help="Notes about this user.")
+@click.option(
+    "--archive-size-quota", help="The total size (in bytes) allowed for this user."
+)
+@sdk_options()
+def update_user(
+    state,
+    user_id,
+    username,
+    email,
+    password,
+    first_name,
+    last_name,
+    notes,
+    archive_size_quota,
+):
+    """Update a user with the specified unique identifier."""
+    _update_user(
+        state.sdk,
+        user_id,
+        username,
+        email,
+        password,
+        first_name,
+        last_name,
+        notes,
+        archive_size_quota,
+    )
+
+
+_bulk_user_update_headers = [
+    "user_id",
+    "username",
+    "email",
+    "password",
+    "first_name",
+    "last_name",
+    "notes",
+    "archive_size_quota",
+]
+
+
+@users.group(cls=OrderedGroup)
+@sdk_options(hidden=True)
+def bulk(state):
+    """Tools for managing users in bulk"""
+    pass
+
+
+users_generate_template = generate_template_cmd_factory(
+    group_name="users",
+    commands_dict={"update": _bulk_user_update_headers},
+    help_message="Generate the CSV template needed for bulk user commands.",
+)
+bulk.add_command(users_generate_template)
+
+
+@bulk.command(name="update")
+@read_csv_arg(headers=_bulk_user_update_headers)
+@format_option
+@sdk_options()
+def bulk_update(state, csv_rows, format):
+    """Update a list of users from the provided CSV."""
+    csv_rows[0]["updated"] = "False"
+    formatter = OutputFormatter(format, {key: key for key in csv_rows[0].keys()})
+
+    def handle_row(**row):
+        try:
+            _update_user(
+                state.sdk, **{key: row[key] for key in row.keys() if key != "updated"}
+            )
+            row["updated"] = "True"
+        except Exception as err:
+            row["updated"] = f"False: {err}"
+        return row
+
+    result_rows = run_bulk_process(
+        handle_row, csv_rows, progress_label="Updating users:"
+    )
+    formatter.echo_formatted_list(result_rows)
+
+
 def _add_user_role(sdk, username, role_name):
     user_id = _get_user_id(sdk, username)
     _get_role_id(sdk, role_name)  # function provides role name validation
@@ -122,3 +220,26 @@ def _get_users_dataframe(sdk, columns, org_uid, role_id, active):
         users_list.extend(page["users"])
 
     return DataFrame.from_records(users_list, columns=columns)
+
+
+def _update_user(
+    sdk,
+    user_id,
+    username,
+    email,
+    password,
+    first_name,
+    last_name,
+    notes,
+    archive_size_quota,
+):
+    return sdk.users.update_user(
+        user_id,
+        username=username,
+        email=email,
+        password=password,
+        first_name=first_name,
+        last_name=last_name,
+        notes=notes,
+        archive_size_quota_bytes=archive_size_quota,
+    )